feat(compute,api,update): Add system version to /health, allow resin pull
This commit adds an environment variable in the 3.3.0 dockerfile that should be maintained as a
semver of the system environment. This is sent in the /health endpoints of both the api server and
update server:

GET /health
200 OK
{
    "name": "opentrons-moon-moon",
    "api_version": "3.3.0-beta.1",
    "fw_version": "edge-6168d32",
    "logs": [
        "/logs/serial.log",
        "/logs/api.log"
    ],
    "system_version": "unknown"
}

GET /server/update/health
200 OK
{
    "name": "opentrons-moon-moon",
    "updateServerVersion": "3.3.0-beta.1",
    "apiServerVersion": "3.3.0-beta.1",
    "smoothieVersion": "edge-6168d32",
    "systemVersion": "3.3.0+resin"
}

In addition, this commit removes the resin-updates.lock file that prevents the host
OS from pulling container changes.